# YJIT: Spill/load argument registers to reuse blocks