This PR improves the argument register allocator in `gen_send_iseq` to reuse existing blocks with different registers.  It previously allowed only moving operands already in a register, but this PR also allows spilling and loading registers. It allows YJIT to avoid compiling duplicated blocks to handle different `RegMapping` variations in most cases.

## YJIT stats
On lobsters, it reduces compilation time and inline_code_size by 3.3% (48ms and 154KB).

### Before
```
compiled_block_count:   71,284
inline_code_size:    4,728,964
outlined_code_size:  4,404,724
code_region_size:   11,628,544
yjit_alloc_size:    26,731,381
yjit_compile_time:   1464.60ms
```

### After
```
compiled_block_count:     68,339
inline_code_size:      4,574,334
outlined_code_size:    4,060,763
code_region_size:     11,157,504
yjit_alloc_size:      25,974,041
yjit_compile_time:     1416.86ms
```

## Benchmark
It seems to have a positive impact on multiple benchmarks. `chunky-png`, `erubi-rails`, `liquid-compile`, and `ruby-lsp` show 4-5% improvements. `activerecord`, `hexapdf`, and `psych-load` are 1% better too.

```
before: ruby 3.4.0dev (2024-12-07T10:02:17Z master bd831bcca5) +YJIT +PRISM [x86_64-linux]
after: ruby 3.4.0dev (2024-12-08T09:06:52Z yjit-reg-reuse 91c1d394f7) +YJIT +PRISM [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  -------------  ------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
activerecord    97.3         0.3         95.8        0.2         1.034          1.016
chunky-png      250.0        0.4         239.1       0.3         1.043          1.046
erubi-rails     486.8        0.2         466.6       0.2         1.042          1.043
hexapdf         1040.9       2.1         1022.1      2.0         1.092          1.018
liquid-c        30.4         0.9         30.5        0.9         1.026          0.997
liquid-compile  34.1         0.8         32.6        2.7         1.021          1.047
liquid-render   40.8         0.6         41.1        0.6         1.021          0.991
lobsters        510.8        2.0         507.2       1.5         1.016          1.007
mail            70.4         0.3         70.4        0.3         1.022          1.000
psych-load      972.9        0.1         957.1       0.1         1.023          1.017
railsbench      1200.5       1.5         1197.6      2.7         1.037          1.002
rubocop         61.9         4.4         61.4        4.1         1.031          1.007
ruby-lsp        72.4         0.9         68.6        0.7         1.018          1.056
sequel          34.0         0.4         33.8        0.4         0.999          1.005
--------------  -----------  ----------  ----------  ----------  -------------  ------------
```