@XrXr 2024-12-09
> It allows YJIT to avoid compiling duplicated blocks to handle different RegMapping variations in most cases.

Curious in what cases the attempt to reuse fail?