@k0kubun 2024-12-09
> Curious in what cases the attempt to reuse fail?

First, when an ISEQ has been compiled as a JIT entry and then compiled as a JIT callsite, it fails. The entry block would have no registers allocated after the prologue code, but the JIT callsite may have some operands in registers. They use a different set of registers, so just shuffling registers wasn't enough. An easy example is `fib`.

Otherwise, it also easily fails when an ISEQ is called from multiple blocks. It can be multiple different caller ISEQs, which may spill at different times, and it can also be the above (entry vs non-entry) problem propagated to callees (when the caller block is simple enough to have no spill prior to the call).